Crate rust_ext

Modules

algebra

block_structure

chain_complex

change_of_basis

cli_module_loaders

combinatorics

fp_vector
An FpVector is a vector with entries in Fp. We use this instead of Vec<u32> because we can pack a lot of entries into a single u64, especially for p small. This not only saves memory, but also leads to faster addition, for example (e.g. a single ^ can add 64 elements of F2 at the same time).

matrix

module

resolution

resolution_homomorphism

steenrod_evaluator

steenrod_parser

yoneda

Structs

AlgebraicObjectsBundle

Config

ModuleFileNotFoundError

Functions

construct

construct_derived_resolution

construct_from_json

load_module_from_file

run_define_module

run_resolve

run_test

run_yoneda

Type Definitions

CCC