Trait rust_ext::chain_complex::ChainComplex

pub trait ChainComplex {
    type Module: Module;
    type Homomorphism: ModuleHomomorphism<Source = Self::Module, Target = Self::Module>;
    fn algebra(&self) -> Arc<AlgebraAny>;

    fn min_degree(&self) -> i32;

    fn zero_module(&self) -> Arc<Self::Module>;

    fn module(&self, homological_degree: u32) -> Arc<Self::Module>;

    fn differential(&self, homological_degree: u32) -> Arc<Self::Homomorphism>;

    fn compute_through_bidegree(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    );

    fn set_homology_basis(
        &self, 
        homological_degree: u32, 
        internal_degree: i32, 
        homology_basis: Vec<usize>
    );

    fn homology_basis(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) -> &Vec<usize>;

    fn max_homology_degree(&self, homological_degree: u32) -> i32;

    fn prime(&self) -> u32 { ... }

    fn compute_homology_through_bidegree(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) { ... }

    fn homology_dimension(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) -> usize { ... }

    fn homology_gen_to_cocyle(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        homological_degree: u32, 
        internal_degree: i32, 
        index: usize
    ) { ... }

    fn compute_homology(&self, homological_degree: u32, internal_degree: i32) { ... }
}

A chain complex is defined to start in degree 0. The min_degree is the min_degree of the modules in the chain complex, all of which must be the same.

Associated Types

type Module: Module

type Homomorphism: ModuleHomomorphism<Source = Self::Module, Target = Self::Module>

Loading content...
Required methods

fn algebra(&self) -> Arc<AlgebraAny>

fn min_degree(&self) -> i32

fn zero_module(&self) -> Arc<Self::Module>

fn module(&self, homological_degree: u32) -> Arc<Self::Module>

fn differential(&self, homological_degree: u32) -> Arc<Self::Homomorphism>

fn compute_through_bidegree(
&self,
homological_degree: u32,
internal_degree: i32
)

fn set_homology_basis(
&self,
homological_degree: u32,
internal_degree: i32,
homology_basis: Vec<usize>
)

fn homology_basis(
&self,
homological_degree: u32,
internal_degree: i32
) -> &Vec<usize>

fn max_homology_degree(&self, homological_degree: u32) -> i32

Loading content...
Provided methods

fn prime(&self) -> u32

fn compute_homology_through_bidegree(
&self,
homological_degree: u32,
internal_degree: i32
)

fn homology_dimension(
&self,
homological_degree: u32,
internal_degree: i32
) -> usize

fn homology_gen_to_cocyle(
&self,
result: &mut FpVector,
coeff: u32,
homological_degree: u32,
internal_degree: i32,
index: usize
)

fn compute_homology(&self, homological_degree: u32, internal_degree: i32)

Loading content...