Struct rust_ext::matrix::Matrix

pub struct Matrix {
    p: u32,
    rows: usize,
    columns: usize,
    slice_row_start: usize,
    slice_row_end: usize,
    slice_col_start: usize,
    slice_col_end: usize,
    vectors: Vec<FpVector>,
}

A matrix! In particular, a matrix with values in F_p. The way we store matrices means it is easier to perform row operations than column operations, and the way we use matrices means we want our matrices to act on the right. Hence we think of vectors as row vectors.

Matrices can be sliced, i.e. restricted to a sub-matrix, and a sliced matrix behaves as if the other rows and columns are not present for many purposes. For example, this affects the values of M[i], the rows and columns functions, as well as more "useful" functions like row_reduce and compute_kernel. However, the row slicing is not taken into account when dereferencing into &[FpVector] (even though the FpVectors still remember the column slicing). This may or may not be a bug.

In general, before one uses a matrix, they must run fp_vector::initialize_limb_bit_index_table(p). This only has to be done once and will be omitted from all examples.

Fields
p: u32rows: usizecolumns: usizeslice_row_start: usizeslice_row_end: usizeslice_col_start: usizeslice_col_end: usizevectors: Vec<FpVector>
Methods

impl Matrix

pub fn new(p: u32, rows: usize, columns: usize) -> Matrix

Produces a new matrix over F_p with the specified number of rows and columns, intiialized to the 0 matrix.

pub fn from_rows(p: u32, vectors: Vec<FpVector>, columns: usize) -> Self

Produces a matrix from a list of rows. If vectors.len() == 0, this returns a matrix with 0 rows and columns. The function does not check if the rows have the same length, but please only input rows that do have the same length.

pub fn from_vec(p: u32, input: &[Vec<u32>]) -> Matrix

Produces a Matrix from an &[Vec<u32>] object. If the number of rows is 0, the number of columns is also assumed to be zero.

Example

let p = 7;
let input  = [vec![1, 3, 6],
              vec![0, 3, 4]];

let m = Matrix::from_vec(p, &input);

pub fn to_vec(&self) -> Vec<Vec<u32>>

pub fn augmented_from_vec(p: u32, input: &[Vec<u32>]) -> (usize, Matrix)

Produces a padded augmented matrix from an &[Vec<u32>] object (produces [A|0|I] from A). Returns the matrix and the first column index of I.

Example

let p = 7;
let input  = [vec![1, 3, 6],
              vec![0, 3, 4]];

let (n, m) = Matrix::augmented_from_vec(p, &input);
assert_eq!(n, FpVector::padded_dimension(p, input[0].len()));

pub fn set_identity(&mut self, size: usize, row: usize, column: usize)

pub fn prime(&self) -> u32

pub fn rows(&self) -> usize

Gets the number of rows in the matrix.

pub fn columns(&self) -> usize

Gets the number of columns in the matrix.

pub fn set_slice(
&mut self,
row_start: usize,
row_end: usize,
col_start: usize,
col_end: usize
)

Sets the slice on the matrix. Restricts to the submatrix consisting of the rows from row_start up to but not including row_end, and the columns from col_start up to but not including col_end.

Slicing modifies the matrix in place.

Example

let p = 3;
let input  = [vec![1, 2, 1, 1, 0],
              vec![1, 0, 2, 1, 0],
              vec![0, 1, 0, 2, 0]];

let mut m = Matrix::from_vec(p, &input);
m.set_slice(1, 4, 1, 3);

assert_eq!(m.rows(), 3);
assert_eq!(m.columns(), 2);
assert_eq!(m[0].entry(0), 0);

pub fn clear_slice(&mut self)

Un-slice the matrix.

pub fn into_slice(self) -> Self

pub fn into_vec(self) -> Vec<FpVector>

impl Matrix

pub fn iter(&self) -> Iter<FpVector>

pub fn iter_mut(&mut self) -> IterMut<FpVector>

impl Matrix

pub fn swap_rows(&mut self, i: usize, j: usize)

pub fn row_op(&mut self, target: usize, source: usize, coeff: u32)

pub fn row_reduce(&mut self, column_to_pivot_row: &mut Vec<isize>)

Perform row reduction to reduce it to reduced row echelon form. This modifies the matrix in place and records the pivots in column_to_pivot_row. The way the pivots are recorded is that column_to_pivot_row[i] is the row of the pivot if the ith row contains a pivot, and -1 otherwise.

One has to call fp_vector::initialize_limb_bit_index_table(p). This step will be skipped in future examples.

Arguments

column_to_pivot_row - A vector for the function to write the pivots into. The length should be at least as long as the number of columns (and the extra entries are ignored).

Example

let p = 7;

let input  = [vec![1, 3, 6],
              vec![0, 3, 4]];

let result = [vec![1, 0, 2],
              vec![0, 1, 6]];

let mut m = Matrix::from_vec(p, &input);
let mut output_pivots_cvec = vec![-1; m.columns()];
m.row_reduce(&mut output_pivots_cvec);

assert_eq!(m, Matrix::from_vec(p, &result));

pub fn row_reduce_offset(
&mut self,
column_to_pivot_row: &mut Vec<isize>,
offset: usize
)

pub fn find_pivots_permutation<T: Iterator<Item = usize>>(
&mut self,
permutation: T
) -> Vec<usize>

This is very similar to row_reduce, except we only need to get to row echelon form, not reduced row echelon form. It also returns the list of pivots instead.

pub fn row_reduce_permutation<T>(
&mut self,
column_to_pivot_row: &mut Vec<isize>,
permutation: T
) where
T: Iterator<Item = usize>,

impl Matrix

pub fn set_to_zero(&mut self)

pub fn find_first_row_in_block(
&self,
pivots: &[isize],
first_column_in_block: usize
) -> usize

pub fn compute_kernel(
&mut self,
column_to_pivot_row: &[isize],
first_source_column: usize
) -> Subspace

Computes the kernel from an augmented matrix in rref. To compute the kernel of a matrix A, produce an augmented matrix of the form

[A | I]

An important thing to note is that the number of columns of A should be a multiple of the number of entries per limb in an FpVector, and this is often achieved by padding columns with 0. The padded length can be obtained from FpVector::padded_dimension.

After this matrix is set up, perform row reduction with Matrix::row_reduce, and then apply compute_kernel.

Arguments

column_to_pivot_row - This is the list of pivots row_reduce gave you.

first_source_column - The column where the I part of the augmented matrix starts.

Example

let p = 3;
let input  = [vec![1, 2, 1, 1, 0],
              vec![1, 0, 2, 1, 1],
              vec![2, 2, 0, 2, 1]];

let (padded_cols, mut m) = Matrix::augmented_from_vec(p, &input);
let mut pivots = vec![-1; m.columns()];
m.row_reduce(&mut pivots);
let ker = m.compute_kernel(&pivots, padded_cols);

let mut target = vec![0; 3];
ker.matrix[0].unpack(&mut target);
assert_eq!(target, vec![1, 1, 2]);

pub fn compute_quasi_inverse(
&mut self,
pivots: &[isize],
last_target_col: usize,
first_source_col: usize
) -> QuasiInverse

Computes the quasi-inverse of a matrix given a rref of [A|0|I], where 0 is the zero padding as usual.

Arguments

pivots - Pivots returned by row_reduce

last_target_col - the last column of A

first_source_col - the first column of I

Example

let p = 3;
let input  = [vec![1, 2, 1, 1, 0],
              vec![1, 0, 2, 1, 1],
              vec![2, 2, 0, 2, 1]];

let (padded_cols, mut m) = Matrix::augmented_from_vec(p, &input);
let mut pivots = vec![-1; m.columns()];
m.row_reduce(&mut pivots);
let qi = m.compute_quasi_inverse(&pivots, input[0].len(), padded_cols);

let image = [vec![1, 0, 2, 1, 1],
             vec![0, 1, 1, 0, 1]];
let computed_image = qi.image.unwrap();
assert_eq!(computed_image.matrix, Matrix::from_vec(p, &image));
assert_eq!(computed_image.column_to_pivot_row, vec![0, 1, -1, -1, -1]);

let preimage = [vec![0, 1, 0],
                vec![0, 2, 2]];
assert_eq!(qi.preimage, Matrix::from_vec(p, &preimage));

pub fn compute_quasi_inverses(
&mut self,
pivots: &[isize],
first_res_col: usize,
last_res_col: usize,
first_source_col: usize
) -> (QuasiInverse, QuasiInverse)

This function computes quasi-inverses for matrices A, B given a reduced row echelon form of [A|0|B|0|I] such that the [A|0] and [B|0] blocks have number of columns a multiple of entries_per_64_bit, and A is surjective. Moreover, if Q is the quasi-inverse of A, it is guaranteed that the image of QB and B|_{ker A} are disjoint.

Arguments

pivots - the pivots produced by row_reduce

first_res_column - the first column of B

last_res_col - the last column of B

first_source_col - the first column of I

pub fn get_image(
&mut self,
image_rows: usize,
target_dimension: usize,
pivots: &[isize]
) -> Subspace

pub fn extend_to_surjection(
&mut self,
first_empty_row: usize,
start_column: usize,
end_column: usize,
current_pivots: &[isize]
) -> Vec<usize>

Given a matrix M in rref, add rows to make the matrix surjective when restricted to the columns between start_column and end_column. That is, if M = [|B|] where B is between columns start_column and end_column, we want the new B to be surjective. This doesn't change the size of the matrix. Rather, it adds the new row to the next empty row in the matrix. This will panic if there are not enough empty rows.

The rows added are all zero except in a single column, where it is 1. The function returns the list of such columns.

Arguments

first_empty_row - The first row in the matrix that is empty. This is where we will add our new rows. This is a mutable borrow and by the end of the function, first_empty_row will be updated to the new first empty row.

current_pivots - The current pivots of the matrix.

Panics

The function panics if there are not enough empty rows.

pub fn extend_image_to_desired_image(
&mut self,
first_empty_row: usize,
start_column: usize,
end_column: usize,
current_pivots: &[isize],
desired_image: &Subspace
) -> Vec<usize>

Given a matrix in rref, say [A|B|C], where B lies between columns start_column and end_columns, and a subspace of the image of B, add rows to the matrix such that the image of B becomes this subspace. This doesn't change the size of the matrix. Rather, it adds the new row to the next empty row in the matrix. This will panic if there are not enough empty rows.

The rows added are basis vectors of the desired image as specified in the Subspace object. The function returns the list of new pivot columns.

Arguments

first_empty_row - The first row in the matrix that is empty. This is where we will add our new rows. This is a mutable borrow and by the end of the function, first_empty_row will be updated to the new first empty row.

current_pivots - The current pivots of the matrix.

Panics

The function panics if there are not enough empty rows. It may panic if the current image is not contained in desired_image, but is not guaranteed to do so.

pub fn extend_image(
&mut self,
first_empty_row: usize,
start_column: usize,
end_column: usize,
current_pivots: &[isize],
desired_image: &Option<Subspace>
) -> Vec<usize>

Extends the image of a matrix to either the whole codomain, or the desired image specified by desired_image. It simply calls extends_image_to_surjection or extend_image_to_surjection depending on the value of desired_image. Refer to these functions for documentation.

pub fn apply(&self, result: &mut FpVector, coeff: u32, input: &FpVector)

Applies a matrix to a vector.

Example

let p = 7;
let input  = [vec![1, 3, 6],
              vec![0, 3, 4]];

let m = Matrix::from_vec(p, &input);
let mut v = FpVector::new(p, 2);
v.pack(&vec![3, 1]);
let mut result = FpVector::new(p, 3);
let mut desired_result = FpVector::new(p, 3);
desired_result.pack(&vec![3, 5, 1]);
m.apply(&mut result, 1, &v);
assert_eq!(result, desired_result);