Struct rust_ext::matrix::Subspace

pub struct Subspace {
    pub matrix: Matrix,
    pub column_to_pivot_row: Vec<isize>,
}

A subspace of a vector space.

Fields

matrix - A matrix in reduced row echelon, whose number of columns is the dimension of the ambient space and each row is a basis vector of the subspace.

column_to_pivot_row - If the column is a pivot column, the entry is the row the pivot corresponds to. If the column is not a pivot column, this is some negative number — not necessarily -1!

Fields
matrix: Matrixcolumn_to_pivot_row: Vec<isize>
Methods

impl Subspace

pub fn new(p: u32, rows: usize, columns: usize) -> Self

pub fn subquotient(
space: Option<&Subspace>,
subspace: Option<&Subspace>,
ambient_dimension: usize
) -> Vec<usize>

Given a chain of subspaces subspace < space < k^ambient_dimension, compute the subquotient space/subspace. The answer is expressed as a list of basis vectors of space whose image in space/subspace forms a basis, and a basis vector of space is described by its index in the list of basis vectors of space (not the ambient space).

Arguments

space - If this is None, it is the whole space k^ambient_dimension

subspace - If this is None, it is empty

pub fn entire_space(p: u32, dim: usize) -> Self

pub fn add_vector(&mut self, row: &FpVector)

This adds a vector to the subspace. This function assumes that the last row of the matrix is zero, i.e. the dimension of the current subspace is strictly less than the number of rows. This can be achieved by setting the number of rows to be the dimension plus one when creating the subspace.

pub fn add_vectors(&mut self, rows: impl Iterator<Item = FpVector>)

pub fn add_basis_elements(&mut self, rows: impl Iterator<Item = usize>)

pub fn reduce(&self, vector: &mut FpVector)

Projects a vector to a complement of the subspace. The complement is the set of vectors that have a 0 in every column where there is a pivot in matrix

pub fn shift_reduce(&self, vector: &mut FpVector)

A version of reduce that doesn't require the vectors to be aligned.

pub fn row_reduce(&mut self)

pub fn contains(&self, vector: &FpVector) -> bool

pub fn dimension(&self) -> usize

pub fn basis(&self) -> &[FpVector]

Returns a basis of the subspace.

pub fn set_to_zero(&mut self)

Sets the subspace to be the zero subspace.

pub fn set_to_entire(&mut self)

Sets the subspace to be the entire subspace.