Struct rust_ext::module::free_module::FreeModule

pub struct FreeModule {
    pub algebra: Arc<AlgebraAny>,
    pub name: String,
    pub min_degree: i32,
    lock: Mutex<()>,
    gen_names: OnceBiVec<Vec<String>>,
    gen_deg_idx_to_internal_idx: OnceBiVec<usize>,
    pub table: OnceBiVec<FreeModuleTableEntry>,
}

Fields
algebra: Arc<AlgebraAny>name: Stringmin_degree: i32lock: Mutex<()>gen_names: OnceBiVec<Vec<String>>gen_deg_idx_to_internal_idx: OnceBiVec<usize>table: OnceBiVec<FreeModuleTableEntry>
Methods

impl FreeModule

pub fn new(algebra: Arc<AlgebraAny>, name: String, min_degree: i32) -> Self

pub fn lock(&self) -> MutexGuard<()>

pub fn max_computed_degree(&self) -> i32

pub fn number_of_gens_in_degree(&self, degree: i32) -> usize

pub fn construct_table(&self, degree: i32) -> FreeModuleTableEntry

pub fn dimension_with_table(table: &FreeModuleTableEntry) -> usize

pub fn add_generators(
&self,
degree: i32,
lock: &MutexGuard<()>,
table: FreeModuleTableEntry,
num_gens: usize,
names: Option<Vec<String>>
)

fn add_generators_to_table(
&self,
degree: i32,
table: &mut FreeModuleTableEntry,
num_gens: usize
)

pub fn generator_offset(
&self,
degree: i32,
gen_deg: i32,
gen_idx: usize
) -> usize

pub fn operation_generator_to_index(
&self,
op_deg: i32,
op_idx: usize,
gen_deg: i32,
gen_idx: usize
) -> usize

pub fn operation_generator_pair_to_idx(
&self,
op_gen: &OperationGeneratorPair
) -> usize

pub fn index_to_op_gen(
&self,
degree: i32,
index: usize
) -> &OperationGeneratorPair

pub fn element_to_json(&self, degree: i32, elt: &FpVector) -> Value

pub fn add_generators_immediate(
&self,
degree: i32,
num_gens: usize,
gen_names: Option<Vec<String>>
)

pub fn add_num_generators(
&self,
degree: i32,
lock: &MutexGuard<()>,
num_gens: usize,
gen_names: Option<Vec<String>>
)

pub fn extend_by_zero(&self, degree: i32)

pub fn get_degrees_with_gens(&self, max_degree: i32) -> Vec<i32>

impl FreeModule

fn standard_act(
&self,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op_index: usize,
input_degree: i32,
input: &FpVector
)

fn custom_milnor_act(
&self,
algebra: &MilnorAlgebra,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op_index: usize,
input_degree: i32,
input: &FpVector
)

For the Milnor algebra, there is a faster algorithm for computing the action, which I learnt from Christian Nassau. This is only implemented for p = 2 for now.

To compute $\mathrm{Sq}(R) \mathrm{Sq}(S)$, we need to iterate over all admissible matrices, namely the $x_{i, j}$ such that $r_i = \sum x_{i, j} p^j$ and the column sums are the $s_j$.

Now if we want to compute $\mathrm{Sq}(R) (\mathrm{Sq}(S^{(1)}) + \cdots)$, we can omit the 0th row of the matrix and the column sum condition. The such a matrix $x_{i, j}$ contributes to $\mathrm{Sq}(R) \mathrm{Sq}(S^{(k)})$ iff the column sum is at most $s_{j}^{(k)}$. There are also some bitwise disjointness conditions we have to check to ensure the coefficient is non-zero.