Struct rust_ext::module::homomorphism::BoundedModuleHomomorphism

pub struct BoundedModuleHomomorphism<S: BoundedModule, T: Module> {
    pub lock: Mutex<()>,
    pub source: Arc<S>,
    pub target: Arc<T>,
    pub degree_shift: i32,
    pub matrices: BiVec<Matrix>,
    pub quasi_inverses: OnceBiVec<QuasiInverse>,
    pub kernels: OnceBiVec<Subspace>,
}

Fields
lock: Mutex<()>source: Arc<S>target: Arc<T>degree_shift: i32matrices: BiVec<Matrix>quasi_inverses: OnceBiVec<QuasiInverse>kernels: OnceBiVec<Subspace>
Methods

impl<S: BoundedModule, T: Module> BoundedModuleHomomorphism<S, T>

pub fn from<F: ModuleHomomorphism<Source = S, Target = T>>(f: &F) -> Self

pub fn replace_source<S_: BoundedModule>(
self,
source: Arc<S_>
) -> BoundedModuleHomomorphism<S_, T>

This function replaces the source of the BoundedModuleHomomorphism and does nothing else. This is useful for changing the type of the source (but not the mathematical module itself). This is intended to be used in conjunction with BoundedModule::to_fd_module

pub fn replace_target<T_: BoundedModule>(
self,
target: Arc<T_>
) -> BoundedModuleHomomorphism<S, T_>

See replace_source