Trait rust_ext::module::homomorphism::ModuleHomomorphism

pub trait ModuleHomomorphism {
    type Source: Module;
    type Target: Module;
    fn source(&self) -> Arc<Self::Source>;

    fn target(&self) -> Arc<Self::Target>;

    fn degree_shift(&self) -> i32;

    fn apply_to_basis_element(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        input_degree: i32, 
        input_idx: usize
    );

    fn kernel(&self, degree: i32) -> &Subspace;

    fn quasi_inverse(&self, degree: i32) -> &QuasiInverse;

    fn compute_kernels_and_quasi_inverses_through_degree(&self, degree: i32);

    fn apply(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        input_degree: i32, 
        input: &FpVector
    ) { ... }

    fn prime(&self) -> u32 { ... }

    fn min_degree(&self) -> i32 { ... }

    fn image(&self, degree: i32) -> &Option<Subspace> { ... }

    fn kernel_and_quasi_inverse(&self, degree: i32) -> (Subspace, QuasiInverse) { ... }

    fn get_matrix(
        &self, 
        matrix: &mut Matrix, 
        degree: i32, 
        start_row: usize, 
        start_column: usize
    ) -> (usize, usize) { ... }
}

Associated Types

type Source: Module

type Target: Module

Loading content...
Required methods

fn source(&self) -> Arc<Self::Source>

fn target(&self) -> Arc<Self::Target>

fn degree_shift(&self) -> i32

fn apply_to_basis_element(
&self,
result: &mut FpVector,
coeff: u32,
input_degree: i32,
input_idx: usize
)

Calling this function when input_idx < source().dimension(input_degree) results in undefined behaviour. Implementations are encouraged to panic when this happens (this is usually the case because of out-of-bounds errors.

fn kernel(&self, degree: i32) -> &Subspace

fn quasi_inverse(&self, degree: i32) -> &QuasiInverse

fn compute_kernels_and_quasi_inverses_through_degree(&self, degree: i32)

Loading content...
Provided methods

fn apply(
&self,
result: &mut FpVector,
coeff: u32,
input_degree: i32,
input: &FpVector
)

fn prime(&self) -> u32

fn min_degree(&self) -> i32

fn image(&self, degree: i32) -> &Option<Subspace>

Returns the image of the module homomorphism in degree degree. If None, the image is the whole space.

fn kernel_and_quasi_inverse(&self, degree: i32) -> (Subspace, QuasiInverse)

fn get_matrix(
&self,
matrix: &mut Matrix,
degree: i32,
start_row: usize,
start_column: usize
) -> (usize, usize)

Loading content...