Struct rust_ext::module::quotient_module::QuotientModule

pub struct QuotientModule<M: Module> {
    pub module: Arc<M>,
    pub subspaces: OnceBiVec<Subspace>,
    pub basis_list: OnceBiVec<Vec<usize>>,
}

Given a module module, this is the quotient of module by a collection of basis elements.

Fields

module - The original module

basis - For each degree d, basis[d] is the list of basis elements of module that are not quotiented out

mask - This is the mask that corresponds to basis. Applying the mask will project to the subspace defined by basis.

Fields
module: Arc<M>subspaces: OnceBiVec<Subspace>basis_list: OnceBiVec<Vec<usize>>
Methods

impl<M: Module> QuotientModule<M>

pub fn new(module: Arc<M>) -> Self

pub fn quotient(&mut self, degree: i32, element: &FpVector)

pub fn quotient_basis_elements(
&mut self,
degree: i32,
elements: impl Iterator<Item = usize>
)

pub fn quotient_vectors(&mut self, degree: i32, vecs: Vec<FpVector>)

fn flush(&mut self, degree: i32)

pub fn quotient_all(&mut self, degree: i32)

pub fn act_on_original_basis(
&self,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op_index: usize,
mod_degree: i32,
mod_index: usize
)

pub fn reduce(&self, degree: i32, vec: &mut FpVector)

pub fn old_basis_to_new(&self, degree: i32, new: &mut FpVector, old: &FpVector)