Struct rust_ext::resolution::Resolution

pub struct Resolution<M: Module, F: ModuleHomomorphism<M, M>, CC: ChainComplex<M, F>> {
    self_: Option<Weak<RefCell<Resolution<M, F, CC>>>>,
    complex: Rc<CC>,
    modules: OnceVec<Rc<FreeModule>>,
    zero_module: Rc<FreeModule>,
    chain_maps: OnceVec<FreeModuleHomomorphism<M>>,
    differentials: OnceVec<FreeModuleHomomorphism<FreeModule>>,
    phantom: PhantomData<dyn ChainComplex<M, F>>,
    kernels: OnceBiVec<RefCell<Option<Subspace>>>,
    next_s: Mutex<u32>,
    next_t: Mutex<i32>,
    pub add_class: Option<Box<dyn Fn(u32, i32, &str)>>,
    pub add_structline: Option<Box<dyn Fn(&str, u32, i32, usize, u32, i32, usize)>>,
    pub unit_resolution: Option<Rc<RefCell<ModuleResolution<FiniteModule>>>>,
    product_list: Vec<Cocycle>,
    chain_maps_to_unit_resolution: OnceVec<OnceBiVec<OnceVec<ResolutionHomomorphismToUnit<M, F, CC>>>>,
    max_product_homological_degree: u32,
    self_maps: Vec<SelfMap<M, F, CC>>,
}

#Fields

kernels - For each internal degree, store the kernel of the most recently calculated chain map as returned by generate_old_kernel_and_compute_new_kernel, to be used if we run resolve_through_degree again.

self_ - This should be a weak reference to yourself. This can be set in the method set_self after constructing the resolution, and is automatically taken care of by the construct function in lib. This is useful because we need a weak reference to self when construction resolution homomorphisms to unit and self maps.

Fields
self_: Option<Weak<RefCell<Resolution<M, F, CC>>>>complex: Rc<CC>modules: OnceVec<Rc<FreeModule>>zero_module: Rc<FreeModule>chain_maps: OnceVec<FreeModuleHomomorphism<M>>differentials: OnceVec<FreeModuleHomomorphism<FreeModule>>phantom: PhantomData<dyn ChainComplex<M, F>>kernels: OnceBiVec<RefCell<Option<Subspace>>>next_s: Mutex<u32>next_t: Mutex<i32>add_class: Option<Box<dyn Fn(u32, i32, &str)>>add_structline: Option<Box<dyn Fn(&str, u32, i32, usize, u32, i32, usize)>>unit_resolution: Option<Rc<RefCell<ModuleResolution<FiniteModule>>>>product_list: Vec<Cocycle>chain_maps_to_unit_resolution: OnceVec<OnceBiVec<OnceVec<ResolutionHomomorphismToUnit<M, F, CC>>>>max_product_homological_degree: u32self_maps: Vec<SelfMap<M, F, CC>>
Methods

impl<M: Module, F: ModuleHomomorphism<M, M>, CC: ChainComplex<M, F>> Resolution<M, F, CC>

pub fn new(
complex: Rc<CC>,
add_class: Option<Box<dyn Fn(u32, i32, &str)>>,
add_structline: Option<Box<dyn Fn(&str, u32, i32, usize, u32, i32, usize)>>
) -> Self

pub fn get_max_degree(&self) -> i32

pub fn get_max_hom_deg(&self) -> u32

pub fn get_complex(&self) -> Rc<CC>

pub fn get_module(&self, homological_degree: u32) -> Rc<FreeModule>

pub fn get_number_of_gens_in_bidegree(
&self,
homological_degree: u32,
internal_degree: i32
) -> usize

pub fn get_chain_map(
&self,
homological_degree: u32
) -> &FreeModuleHomomorphism<M>

pub fn get_cocycle_string(
&self,
hom_deg: u32,
int_deg: i32,
idx: usize
) -> String

pub fn set_self(&mut self, self_: Weak<RefCell<Resolution<M, F, CC>>>)

Sets self.self_. See struct documentation for more about self.self_.

pub fn extend_through_degree(
&self,
next_s: u32,
max_s: u32,
next_t: i32,
max_t: i32
)

This function prepares the Resolution object to perform computations up to the specified s degree. It does not perform any computations by itself. It simply lengthens the OnceVecs modules, chain_maps, etc. to the right length.

pub fn resolve_through_bidegree(&self, max_s: u32, max_t: i32)

pub fn resolve_through_degree(&self, degree: i32)

pub fn step(&self, s: u32, t: i32, old_kernel: Option<Subspace>) -> Subspace

fn compute_filtration_one_products(
&self,
homological_degree: u32,
internal_degree: i32,
source_idx: usize
)

pub fn add_structline(
&self,
name: &str,
source_hom_deg: u32,
source_int_deg: i32,
source_idx: usize,
target_hom_deg: u32,
target_int_deg: i32,
target_idx: usize
)

pub fn generate_old_kernel_and_compute_new_kernel(
&self,
homological_degree: u32,
degree: i32,
old_kernel: Option<Subspace>
) -> Subspace

Call our resolution $X$, and the chain complex to resolve $C$. This is a legitimate resolution if the map $f: X \to C$ induces an isomorphism on homology. This is the same as saying the cofiber is exact. The cofiber is given by the complex

$$ X_s \oplus C_{s+1} \to X_{s-1} \oplus C_s \to X_{s-2} \oplus C_{s-1} \to \cdots $$

where the differentials are given by

$$ \begin{pmatrix} d_X & 0 \\ (-1)^s f & d_C \end{pmatrix} $$

Our method of producing $X_{s, t}$ and the chain maps are as follows. Suppose we have already built the chain map and differential for $X_{s-1, t}$ and $X_{s, t-1}$. Since $X_s$ is a free module, the generators in degree $< t$ gives us a bunch of elements in $X_s$ already, and we know exactly where they get mapped to. Let $T$ be the $\mathbb{F}_p$ vector space generated by these elements. Then we already have a map

$$ T \to X_{s-1, t} \oplus C_{s, t}$$

and we know this hits the kernel of the map

$$ D = X_{s-1, t} \oplus C_{s, t} \to X_{s-2, t} \oplus C_{s-1, t}. $$

What we need to do now is to add generators to $X_{s, t}$ to hit the entirity of this kernel. Note that we don't have to do this. Some of the elements in the kernel might be hit by $C_{s+1, t}$ and we don't have to hit them, but we opt to add generators to hit it anyway.

If we do it this way, then we know the composite of the map

$$ T \to X_{s-1, t} \oplus C_{s, t} \to C_{s, t} $$

has to be surjective, since the image of $C_{s, t}$ under $D$ is also in the image of $X_{s-1, t}$. So our first step is to add generators to $X_{s, t}$ such that this composite is surjective.

After adding these generators, we need to decide where to send them to. We know their values in the $C_{s, t}$ component, but we need to use a quasi-inverse to find the element in $X_{s-1, t}$ that hits the corresponding image of $C_{s, t}$. This tells us the $X_{s-1, t}$ component.

Finally, we need to add further generators to $X_{s, t}$ to hit all the elements in the kernel of

$$ X_{s-1, t} \to X_{s-2, t} \oplus C_{s-1, t}. $$

This kernel was recorded by the previous iteration of the method in old_kernel, so this step is doable as well.

Note that if we add our new generators conservatively, then the kernel of the maps

$$ \begin{aligned} T &\to X_{s-1, t} \oplus C_{s, t} \\ X_{s, t} &\to X_{s-1, t} \oplus C_{s, t} \end{aligned} $$ agree.

In the code, we first row reduce the matrix of the map from $T$. This lets us record the kernel which is what the function returns at the end. This computation helps us perform the future steps since we need to know about the cokernel of this map.

Arguments

homological_degree - The s degree to calculate

degree - The t degree to calculate

old_kernel - The kernel of the map $X_{s-1, t} \to X_{s-2, t} \oplus C_{s-1, t}$, computed and returned by the previous iteration of this function for $(s-1, t)$. This is None when $s = 0$.

pub fn graded_dimension_string(&self) -> String

impl<M, F, CC> Resolution<M, F, CC> where
M: Module,
F: ModuleHomomorphism<M, M>,
CC: ChainComplex<M, F>,

pub fn add_product(&mut self, s: u32, t: i32, index: usize, name: String)

pub fn catch_up_products(&self)

This function computes the products between the element most recently added to product_list and the parts of Ext that have already been computed. This function should be called right after add_product, unless resolve_through_degree/resolve_through_bidegree has never been called.

This is made separate from add_product because extend_maps_to_unit needs a borrow of self, but add_product takes in a mutable borrow.

pub fn construct_unit_resolution(&mut self)

pub fn set_unit_resolution(
&mut self,
unit_res: Rc<RefCell<ModuleResolution<FiniteModule>>>
)

fn compute_products(&self, s: u32, t: i32, products: &[Cocycle])

Compute products whose result lie in degrees up to (s, t)

fn compute_product_step(&self, elt: &Cocycle, s: u32, t: i32, idx: usize)

fn extend_maps_to_unit(&self, s: u32, t: i32)

This ensures the chain_maps_to_unit_resolution are defined such that we can compute products up to bidegree (s, t)

impl<M, F, CC> Resolution<M, F, CC> where
M: Module,
F: ModuleHomomorphism<M, M>,
CC: ChainComplex<M, F>,

pub fn add_self_map(&mut self, s: u32, t: i32, name: String, map_data: Matrix)

fn compute_self_maps(&self, s: u32, t: i32)

We compute the products by self maps where the result has degree (s, t).