Trait rust_ext::fp_vector::FpVectorT

pub trait FpVectorT {
    fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize);

    fn get_vector_container(&self) -> &VectorContainer;

    fn get_vector_container_mut(&mut self) -> &mut VectorContainer;

    fn get_prime(&self) -> u32;

    fn get_dimension(&self) -> usize { ... }

    fn get_offset(&self) -> usize { ... }

    fn get_min_index(&self) -> usize { ... }

    fn get_slice(&self) -> (usize, usize) { ... }

    fn set_slice(&mut self, slice_start: usize, slice_end: usize) { ... }

    fn restore_slice(&mut self, slice: (usize, usize)) { ... }

    fn clear_slice(&mut self) { ... }

    fn get_min_limb(&self) -> usize { ... }

    fn get_max_limb(&self) -> usize { ... }

    fn get_limbs_cvec(&self) -> &Vec<u64> { ... }

    fn get_limbs_cvec_mut(&mut self) -> &mut Vec<u64> { ... }

    fn get_limb_mask(&self, limb_idx: usize) -> u64 { ... }

    fn set_to_zero(&mut self) { ... }

    fn assign(&mut self, other: &FpVector) { ... }

    fn is_zero(&self) -> bool { ... }

    fn get_entry(&self, index: usize) -> u32 { ... }

    fn set_entry(&mut self, index: usize, value: u32) { ... }

    fn add_basis_element(&mut self, index: usize, value: u32) { ... }

    fn unpack(&self, target: &mut [u32]) { ... }

    fn pack(&mut self, source: &[u32]) { ... }

    fn add(&mut self, other: &FpVector, c: u32) { ... }

    fn scale(&mut self, c: u32) { ... }
}

Required methods

fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize)

fn get_vector_container(&self) -> &VectorContainer

fn get_vector_container_mut(&mut self) -> &mut VectorContainer

fn get_prime(&self) -> u32

Loading content...
Provided methods

fn get_dimension(&self) -> usize

fn get_offset(&self) -> usize

fn get_min_index(&self) -> usize

fn get_slice(&self) -> (usize, usize)

fn set_slice(&mut self, slice_start: usize, slice_end: usize)

fn restore_slice(&mut self, slice: (usize, usize))

fn clear_slice(&mut self)

fn get_min_limb(&self) -> usize

fn get_max_limb(&self) -> usize

fn get_limbs_cvec(&self) -> &Vec<u64>

fn get_limbs_cvec_mut(&mut self) -> &mut Vec<u64>

fn get_limb_mask(&self, limb_idx: usize) -> u64

fn set_to_zero(&mut self)

fn assign(&mut self, other: &FpVector)

fn is_zero(&self) -> bool

fn get_entry(&self, index: usize) -> u32

fn set_entry(&mut self, index: usize, value: u32)

fn add_basis_element(&mut self, index: usize, value: u32)

fn unpack(&self, target: &mut [u32])

Unpacks an FpVector onto an array slice. note that the array slice has to be long enough to hold all the elements in the FpVector.

fn pack(&mut self, source: &[u32])

fn add(&mut self, other: &FpVector, c: u32)

fn scale(&mut self, c: u32)

Loading content...