Trait rust_ext::fp_vector::FpVectorT

pub trait FpVectorT {
    fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize);

    fn vector_container(&self) -> &VectorContainer;

    fn vector_container_mut(&mut self) -> &mut VectorContainer;

    fn prime(&self) -> u32;

    fn dimension(&self) -> usize { ... }

    fn offset(&self) -> usize { ... }

    fn min_index(&self) -> usize { ... }

    fn slice(&self) -> (usize, usize) { ... }

    fn set_slice(&mut self, slice_start: usize, slice_end: usize) { ... }

    fn restore_slice(&mut self, slice: (usize, usize)) { ... }

    fn clear_slice(&mut self) { ... }

    fn into_slice(&mut self) { ... }

    fn min_limb(&self) -> usize { ... }

    fn max_limb(&self) -> usize { ... }

    fn limbs(&self) -> &Vec<u64> { ... }

    fn limbs_mut(&mut self) -> &mut Vec<u64> { ... }

    fn limb_mask(&self, limb_idx: usize) -> u64 { ... }

    fn set_to_zero_pure(&mut self) { ... }

    fn set_to_zero(&mut self) { ... }

    fn shift_assign(&mut self, other: &FpVector) { ... }

    fn assign(&mut self, other: &FpVector) { ... }

    fn is_zero_pure(&self) -> bool { ... }

    fn is_zero(&self) -> bool { ... }

    fn entry(&self, index: usize) -> u32 { ... }

    fn set_entry(&mut self, index: usize, value: u32) { ... }

    fn add_basis_element(&mut self, index: usize, value: u32) { ... }

    fn unpack(&self, target: &mut [u32]) { ... }

    fn to_vector(&self) -> Vec<u32> { ... }

    fn pack(&mut self, source: &[u32]) { ... }

    fn add_tensor(
        &mut self, 
        offset: usize, 
        coeff: u32, 
        left: &FpVector, 
        right: &FpVector
    ) { ... }

    fn add(&mut self, other: &FpVector, c: u32) { ... }

    fn shift_add(&mut self, other: &FpVector, c: u32) { ... }

    fn shift_right_add(&mut self, other: &FpVector, c: u32) { ... }

    fn shift_left_add(&mut self, other: &FpVector, c: u32) { ... }

    fn scale(&mut self, c: u32) { ... }
}

Required methods

fn reduce_limbs(&mut self, start_limb: usize, end_limb: usize)

fn vector_container(&self) -> &VectorContainer

fn vector_container_mut(&mut self) -> &mut VectorContainer

fn prime(&self) -> u32

Loading content...
Provided methods

fn dimension(&self) -> usize

fn offset(&self) -> usize

fn min_index(&self) -> usize

fn slice(&self) -> (usize, usize)

fn set_slice(&mut self, slice_start: usize, slice_end: usize)

fn restore_slice(&mut self, slice: (usize, usize))

fn clear_slice(&mut self)

fn into_slice(&mut self)

Drops every element in the fp_vector that is not in the current slice.

fn min_limb(&self) -> usize

fn max_limb(&self) -> usize

fn limbs(&self) -> &Vec<u64>

fn limbs_mut(&mut self) -> &mut Vec<u64>

fn limb_mask(&self, limb_idx: usize) -> u64

fn set_to_zero_pure(&mut self)

fn set_to_zero(&mut self)

fn shift_assign(&mut self, other: &FpVector)

fn assign(&mut self, other: &FpVector)

fn is_zero_pure(&self) -> bool

fn is_zero(&self) -> bool

fn entry(&self, index: usize) -> u32

fn set_entry(&mut self, index: usize, value: u32)

fn add_basis_element(&mut self, index: usize, value: u32)

fn unpack(&self, target: &mut [u32])

Unpacks an FpVector onto an array slice. note that the array slice has to be long enough to hold all the elements in the FpVector.

fn to_vector(&self) -> Vec<u32>

fn pack(&mut self, source: &[u32])

fn add_tensor(
&mut self,
offset: usize,
coeff: u32,
left: &FpVector,
right: &FpVector
)

coeff need not be reduced mod p. Adds v otimes w to self.

fn add(&mut self, other: &FpVector, c: u32)

Adds c * other to self. other must have the same length, offset, and prime as self, and c must be between 0 and p - 1.

fn shift_add(&mut self, other: &FpVector, c: u32)

fn shift_right_add(&mut self, other: &FpVector, c: u32)

fn shift_left_add(&mut self, other: &FpVector, c: u32)

fn scale(&mut self, c: u32)

Loading content...